Add files via upload
The (fictional) village of Hawkshead has had its own library for 25 years, but it is now under threat of closure due to spending cuts. The only hope was to improve the library’s efficiency and to do this they asked me to create a computer program to keep track of its members and books.